Update coldplate3.i
Update coldplate3.i, in this version：
At first, my linear residual convergence was not good, I first try disabling the line search, which slightly made the calculation steps forward a few steps, and then I added adiabatic boundary conditions (previously I thought this was the default in MOOSE ), and later proved that this does not improve the convergence of the calculation (probably MOOSE defaults to adiabatic boundary conditions for unset temperature boundary conditions? I am not sure).

According to Mr.Giud’s suggestion and official guide, I modified the preprocessor code. The linear residual converges very well. This solves the problem of linear residual convergence, but the nonlinear residual still converges poorly, or even at all. There is no convergence.

I also tried to modify the viscosity to 1000, which does improve the continuity of the calculation, and will not prompt non-convergence soon after the start of the calculation, but the nonlinear residuals still have poor convergence.

So according to the official guide, I now turn off the automatic scaling and input the command in the code to print the residuals of each variable. Through debugging, I adjust the variables to the same order of magnitude. The current nonlinear residuals have a magnitude of about 1e-4. , But they often rebound, so it is difficult to achieve convergence.