Create coldplate-small.i
errors:

1 *** ERROR ***
The requested regular material property rho is declared as an AD property. Either retrieve it as an AD property with getADMaterialProperty or declare it as a regular property wtih declareProperty

2.*** ERROR ***
A 'INSADTemperatureTimeDerivative' is not a registered object.

If you are trying to find this object in a dynamically linked library, make sure that
the library can be found either in your "Problem/library_path" parameter or in the
MOOSE_LIBRARY_PATH environment variable.


3.*** ERROR ***
The following error occurred in the object "MOOSE Problem", of type "FEProblem".

No standard variable named velocity found. Did you specify a vector variable when you meant to specify a standard variable?